Refactored simulation setup
